refactor(assemble domain): 관련 도메인 묶음
