+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,9 +1,9 @@
-package kitchenpos.menugroup.application;
+package kitchenpos.menu.application;
 
 import java.util.stream.Collectors;
-import kitchenpos.menugroup.domain.MenuGroupRepository;
-import kitchenpos.menugroup.dto.MenuGroupRequest;
-import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.dto.MenuGroupRequest;
+import kitchenpos.menu.dto.MenuGroupResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
