+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -10,8 +10,8 @@ import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
