+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -5,7 +5,7 @@ import java.util.List;
 import kitchenpos.common.Price;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuRequest {
     private String name;
