+++ src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -1,9 +1,9 @@
-package kitchenpos.menugroup.ui;
+package kitchenpos.menu.ui;
 
 import io.swagger.annotations.ApiOperation;
-import kitchenpos.menugroup.application.MenuGroupService;
-import kitchenpos.menugroup.dto.MenuGroupRequest;
-import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.menu.application.MenuGroupService;
+import kitchenpos.menu.dto.MenuGroupRequest;
+import kitchenpos.menu.dto.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
