+++ src/main/java/kitchenpos/table/domain/OrderTables.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.domain;
+package kitchenpos.table.domain;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,7 +8,6 @@ import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
-import kitchenpos.table.domain.OrderTable;
 
 @Embeddable
 public class OrderTables {
@@ -21,13 +20,6 @@ public class OrderTables {
         OrderTables orderTables = new OrderTables();
         orderTables.addAll(orderTableList);
         return orderTables;
-    }
-
-    public void add(OrderTable orderTable) {
-        if (this.orderTables.contains(orderTable)) {
-            return;
-        }
-        this.orderTables.add(orderTable);
     }
 
     public void addAll(List<OrderTable> orderTables) {
