+++ src/main/java/kitchenpos/table/dto/TableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.dto;
+package kitchenpos.table.dto;
 
 import java.util.List;
 import javax.validation.constraints.NotEmpty;
